First sketch, does not work yet
The difficulty is that we use pkgload and that uses
other packages, so these need to be in the library.
But all dependencies should be removed from the
library as well.

We would need to create a temporary library that only
contains the dependencies we are testing the loading
with. (Plus the packages that `builddeps` needs, but
ideally this should be none. So it would be great to
get rid of the `pkgload` dependency.)